+++ r4300/exception.h
@@ -30,9 +30,11 @@
 #ifndef EXCEPTION_H
 #define EXCEPTION_H
 
+#include "../main/winlnxdefs.h"
+
 void address_error_exception();
 void TLB_invalid_exception();
-void TLB_refill_exception(unsigned long addresse, int w);
+void TLB_refill_exception(u32 addresse, int w);
 void TLB_mod_exception();
 void integer_overflow_exception();
 void coprocessor_unusable_exception();
